move components to components directory
